requirement.txt
certifi==2019.9.11
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-crispy-forms==1.8.0
django-registration==3.0.1
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
django-webpack-loader==0.6.0
djangorestframework==3.10.3
idna==2.8
mysqlclient==1.3.13
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
